ToDo blazor component added

+++ ToDo.Application/ToDoItem/Queries/GetItem/ToDoTaskResponse.cs
@@ -1,5 +1,5 @@
 ﻿namespace ToDo.Application.ToDoItems.Queries.GetToDoItem;
 
-public sealed record ToDoItemResponse(long Id, string Title, bool IsDone);
+public sealed record ToDoItemResponse(int Id, string Title, bool IsDone);
